requirements.txt
# Core dependencies
torch>=2.0.0
torch-geometric>=2.4.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
torch-cluster>=1.6.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torchmetrics>=1.0.0

# NLP dependencies
spacy>=3.7.0
transformers>=4.30.0

# Visualization and demo
matplotlib>=3.7.0
plotly>=5.15.0
streamlit>=1.25.0
networkx>=3.1.0
pyvis>=0.3.2

# Configuration and utilities
pyyaml>=6.0
omegaconf>=2.3.0
tqdm>=4.65.0
wandb>=0.15.0

# Development tools
black>=23.0.0
ruff>=0.1.0
pytest>=7.4.0
pre-commit>=3.3.0

# Optional: for advanced NLP features
nltk>=3.8.0
textblob>=0.17.0
